CSS Update
Logo is now transparent. Background changed to an image of Barrie.  Made
the content divs smaller, and moved them to the left. Removed background
image for the navbar and changed it to a background gradient. Added a
link to a pdf of my resume when the resume image is clicked. Adjusted
image heights and widths as a result of the divs being smaller.